Update TensorRT-LLM (#1554)

+++ windows/setup_env.ps1
@@ -4,7 +4,8 @@ param (
     [switch]$skipPython,
     [switch]$skipMPI = $true,
     [switch]$skipCUDNN,
-    [string]$cudaVersion #CUDA version defaults to 12.3, specify otherwise, however only 12.2 and 12.3 have uris
+    [string]$cudaVersion, #CUDA version defaults to 12.4, specify otherwise
+    [switch]$skipTRT = $true
 )
 
 # Set the error action preference to 'Stop' for the entire script.
@@ -19,13 +20,13 @@ $ErrorActionPreference = 'Stop'
 
 New-Item -Path "$($env:LOCALAPPDATA)\trt_env_outlog.txt" -Force
 
-# Install CUDA, default to 12.3
+# Install CUDA, default to 12.4
 if (-not $skipCUDA){
     if($cudaVersion){
         $cudaVer = "NVIDIA CUDA Toolkit " + $cudaVersion
         } else {
-        $cudaVersion = 12.3
-        $cudaVer = "NVIDIA CUDA Toolkit 12.3"
+        $cudaVersion = 12.4
+        $cudaVer = "NVIDIA CUDA Toolkit 12.4"
     }
 
     if (-not (Get-Package -Name $cudaVer -EA Ignore)) {
@@ -35,6 +36,8 @@ if (-not $skipCUDA){
             Invoke-WebRequest -Uri 'https://developer.download.nvidia.com/compute/cuda/12.2.2/local_installers/cuda_12.2.2_537.13_windows.exe' -OutFile 'cuda_installer.exe'
         } elseif ($cudaVersion -eq 12.3){
             Invoke-WebRequest -Uri 'https://developer.download.nvidia.com/compute/cuda/12.3.2/local_installers/cuda_12.3.2_546.12_windows.exe' -OutFile 'cuda_installer.exe'
+        } elseif ($cudaVersion -eq 12.4){
+            Invoke-WebRequest -Uri 'https://developer.download.nvidia.com/compute/cuda/12.4.0/local_installers/cuda_12.4.0_551.61_windows.exe' -OutFile 'cuda_installer.exe'
         } else {
             $cudaUri = Read-Host "Please go to https://developer.nvidia.com/cuda-downloads and input the url of the CUDA version you wish to use"
             Invoke-WebRequest -Uri $cudaUri -OutFile 'cuda_installer.exe'
@@ -71,7 +74,7 @@ if(-not $skipPython){
         Write-Output "Creating python3 alias executable"
         Copy-Item -Path 'C:\Program Files\Python310\python.exe' -Destination 'C:\Program Files\Python310\python3.exe'
         Write-Output "Done Python installation at 'C:\Program Files\Python310'"
-        [Environment]::SetEnvironmentVariable('Path', "C:\Program Files\Python310;$env:Path", [EnvironmentVariableTarget]::Machine)
+        [Environment]::SetEnvironmentVariable('Path', "C:\Program Files\Python310;C:\Program Files\Python310\Scripts;$env:Path", [EnvironmentVariableTarget]::Machine)
         Add-Content -Path $env:LOCALAPPDATA\trt_env_outlog.txt -Value "0"
     } else {
         Write-Output "Python installation already exists"
@@ -87,6 +90,10 @@ if (-not ($skipMPI)) {
     if (-not (Test-Path -Path 'C:\Program Files\Microsoft MPI\Bin')) {
         Write-Output "Downloading Microsoft MPI not detected"
         Add-Content -Path $env:LOCALAPPDATA\trt_env_outlog.txt -Value "0"
+        # The latest version is 10.1.3, but it requires you to get a temporary download
+        # link.
+        # https://learn.microsoft.com/en-us/message-passing-interface/microsoft-mpi-release-notes
+        # We use 10.1.1 which has a release on the GitHub page
         Write-Output "Downloading Microsoft MPI installer"
         Invoke-WebRequest -Uri 'https://github.com/microsoft/Microsoft-MPI/releases/download/v10.1.1/msmpisetup.exe' -OutFile 'msmpisetup.exe'
         Write-Output "Installing Microsoft MPI"
@@ -162,15 +169,42 @@ if(-not $skipCUDNN){
     Add-Content -Path $env:LOCALAPPDATA\trt_env_outlog.txt -Value "1"
 }
 
-Write-Output "Grabbing TensorRT..."
-$ProgressPreference = 'SilentlyContinue'
-New-Item -Path .\TensorRT -ItemType Directory
-Invoke-WebRequest -Uri 'https://developer.nvidia.com/downloads/compute/machine-learning/tensorrt/9.3.0/tensorrt-9.3.0.1.windows10.win10.cuda-12.2.llm.beta.zip' -OutFile .\TensorRT\trt.zip
-Expand-Archive -Path .\TensorRT\trt.zip -DestinationPath .\TensorRT\Temp
-Move-Item -Path .\TensorRT\Temp\TensorRT-9.3.0.1.Windows10.win10.cuda-12.2.llm.beta\TensorRT-9.3.0.1 -Destination .\TensorRT
-Remove-Item -Path .\TensorRT\trt.zip -Force
-Remove-Item .\TensorRT\Temp -Force -Recurse
-Write-Output "TensorRT installed at .\TensorRT\TensorRT-9.3.0.1"
+# Install TensorRT
+if (-not ($skipTRT)) {
+    $TRT_BASE = Join-Path $PWD \TensorRT
+    if (-not (Test-Path -Path $TRT_BASE)) {
+        Write-Output "Grabbing TensorRT..."
+        $ProgressPreference = 'SilentlyContinue'
+        New-Item -Path .\TensorRT -ItemType Directory
+        Invoke-WebRequest -Uri 'https://developer.nvidia.com/downloads/compute/machine-learning/tensorrt/10.0.1/zip/TensorRT-10.0.1.6.Windows10.win10.cuda-12.4.zip' -OutFile .\TensorRT\trt.zip
+        Expand-Archive -Path .\TensorRT\trt.zip -DestinationPath .\TensorRT\
+        Remove-Item -Path .\TensorRT\trt.zip -Force
+        $trtPath = Join-Path $TRT_BASE TensorRT-10.0.1.6
+        Write-Output "TensorRT installed at ${trtPath}"
+
+        $trtSubPaths = @{
+            "bin" = Join-Path $trtPath bin
+            "include" = Join-Path $trtPath include
+            "lib" = Join-Path $trtPath lib
+        }
+
+        foreach ($key in $trtSubPaths.Keys) {
+            $subPath = $trtSubPaths[$key]
+            if (-not (Test-Path -Path $subPath)) {
+              Write-Error "TensorRT ${key} path ${subPath} does not exist!"
+            }
+        }
+        $TRTEnvVar = $trtSubPaths.Values -join ";"
+
+        [Environment]::SetEnvironmentVariable("TRT", "$TRTEnvVar", [EnvironmentVariableTarget]::Machine)
+    } else {
+        Write-Output "TensorRT already present"
+        Add-Content -Path $env:LOCALAPPDATA\trt_env_outlog.txt -Value "1"
+    }
+} else {
+    Write-Output "Skipping TRT installation"
+    Add-Content -Path $env:LOCALAPPDATA\trt_env_outlog.txt -Value "1"
+}
 
 
 return $env:Path
